requirements.txt
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5
Django==3.0.5
django-crispy-forms==1.9.0
django-filter==2.2.0
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.4.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
wrapt==1.11.2
